moved set_codes functionality to huffman tree
