+++ lab3/huffman_tree.py
@@ -33,6 +33,11 @@ class HuffmanTree:
             [None for _ in range(ord('z') - ord('A') + 1)] if memo else []
         self._root = self._build_tree()
 
+        # Set binary codes for quick retrieval if has memo. Otherwise, find
+        # dynamically during encoding process
+        if memo:
+            self._set_codes()
+
     def __str__(self) -> str:
         """
         String representation of Huffman Tree using pre-order traversal
@@ -147,6 +152,69 @@ class HuffmanTree:
 
         # Last item is the root of a new binary tree. Return root
         return nodes_pq.heap_pop()
+
+    def _set_codes(self) -> 'HuffmanTree':
+        """
+        Method for adding binary codes to Huffman tree nodes.
+
+        Returns:
+            HuffmanTree: current HuffmanTree object instance
+        """
+
+        def preorder(node: Optional['HuffmanNode'], prefix="") -> None:
+            """
+            Helper function for preorder traversal of Huffman Tree to update
+            each node's binary Huffman code. Each left child has a 0 appended,
+            and each right has a 1 appended to its parent node's code (prefix).
+
+            Args:
+                node (HuffmanNode): current node being visited
+                prefix (str): parent node's Huffman code
+            """
+            if not node:
+                return
+
+            if node.is_leaf():
+                node.set_code(prefix)
+            else:
+                preorder(node.get_left(), prefix + "0")
+                preorder(node.get_right(), prefix + "1")
+
+            return
+
+        root = self._root
+        preorder(root)
+
+        return self
+
+    def print_codes(self) -> str:
+        """
+        Method for representing Huffman Encoding binary codes for all Huffman 
+        tree nodes using preorder traversal. Output is a comma-separated list 
+        of each node's characters string and binary Huffman code as a key-value
+        pair.
+
+        str: pre-order traversal of Huffman Tree nodes in format 
+                characters: code
+        """
+        def preorder(root: Optional['HuffmanNode']) -> List[str]:
+            rep = []
+
+            # Visit root, then left, then right
+            if not root:
+                return rep
+
+            rep.append(f"{root.get_characters()}: {root.get_code()}")
+
+            if root.get_left():
+                rep.extend(preorder(root.get_left()))
+
+            if root.get_right():
+                rep.extend(preorder(root.get_right()))
+
+            return rep
+
+        return ', '.join(preorder(self._root))
 
     def get_root(self) -> 'HuffmanNode':
         """
